changed sizes to NSUInteger

+++ SDWebImage/SDImageCache.h
@@ -45,7 +45,7 @@ typedef enum SDImageCacheType SDImageCacheType;
 /**
  * The maximum size of the cache, in bytes.
  */
-@property (assign, nonatomic) unsigned long long maxCacheSize;
+@property (assign, nonatomic) NSUInteger maxCacheSize;
 
 /**
  * Returns global shared cache instance
@@ -153,7 +153,7 @@ typedef enum SDImageCacheType SDImageCacheType;
 /**
  * Get the size used by the disk cache
  */
-- (unsigned long long)getSize;
+- (NSUInteger)getSize;
 
 /**
  * Get the number of images in the disk cache
